New Startup Modes
passing a parameter of top-one-thousand or top-five-thousand will lauch the app and load one of those options right from the start. The main reason to have this is so that the cache for each can be built after the data has been retrieved from TMDB. These are used by the XData server to simulate (or rather force) the request, which then goes through the whole process to fill out the various caches so that when the next person comes along, it is as fast as possible.